Refactor to separate EditorScreen and PlayScreen tests
